restructured files to be more modular
