+++ package.json
@@ -3,11 +3,14 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-16": "^1.1.1",
     "firebase": "^5.0.3",
     "react": "^16.3.2",
     "react-day-picker": "^7.1.9",
     "react-dom": "^16.3.2",
-    "react-scripts": "1.1.4"
+    "react-scripts": "1.1.4",
+    "react-test-renderer": "^16.4.0"
   },
   "scripts": {
     "start": "react-scripts start",
