+++ src/App/App.test.js
@@ -1,0 +1,11 @@
+import React from 'react';
+// import ReactDOM from 'react-dom';
+import App from './App';
+import { shallow } from 'enzyme';
+
+describe('App', () => {
+  it('should match the snapshot', () => {
+    const wrapper = shallow(<App />)
+    expect(wrapper).toMatchSnapshot();
+  })
+})
